% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{AppUI}
\alias{AppUI}
\title{Launch interactive User Interface}
\usage{
AppUI()
}
\arguments{
\item{sfPower}{numerical value of ... for}
}
\value{
A user interface will be shown in users' default web browser.
}
\description{
AppUI initiates in the web browser an interactive user interface of ...  This user interface enables users to easily perform nearly all standard ... functions in ... package.
}
